+++ runner/monitor/CHANGELOG.md
@@ -9,6 +9,9 @@
 * Fixes missing IntentMonitor for startActivity with user overload. Now it's
   possible to intercept intents started with startActivityAsUser.
 
+* Adds missing override of an existing callActivityOnCreate method with
+  PersistableBundle.
+
 **New Features**
 
 **Breaking Changes**